requirements.txt
dj-database-url
Django>=1.11.7,<1.12
djangorestframework
djangorestframework-gis
django-cors-headers
django-compressor
psycopg2
python-decouple
whitenoise
unipath
smartmin
coreapi
coreapi-cli
Pillow
pyyaml
django-resized
django-taggit
gdal
django-filter
django-crispy-forms
django-sass-processor
django-leaflet
numpy
libsass
-e git://github.com/eltonplima/django-qurl-templatetag.git#egg=master
django-docs
python-magic
easy-thumbnails
django-ordered-model